[Ticket #12520] - Unexpected exception warnings in the log when studies are tier-migrated
Surprised this wasn't cleaned up earlier.  The problem was that a cleanup of the
CopyDirectoryCommand.  The Dispose method was attempting to delete a backup
directory that in most cases wasn't set/needed and logging a needless warning
message.   Cleaned up the delete check to make sure that a directory has been
set before attempting to delete, and also cleaned up the logging messages a bit
to make them more helpful when we really do have a problem occurring with the
cleanup.

--HG--
branch : Orion

+++ Dicom/Utilities/Command/CopyDirectoryCommand.cs
@@ -89,16 +89,16 @@ namespace ClearCanvas.Dicom.Utilities.Command
         {
             if (_copied)
             {
-                if (Directory.Exists(_backupDestDir))
+                if (!string.IsNullOrEmpty(_backupDestDir) && Directory.Exists(_backupDestDir))
                 {
                     try
                     {
                         DirectoryUtility.DeleteIfExists(_dest);
                     }
-                    catch
+                    catch (Exception ex)
                     {
                     	// ignore it, will overwrite anyway
-                        Platform.Log(LogLevel.Warn, "Unexpected exeception attempting to delete: {0}", _dest);
+                        Platform.Log(LogLevel.Warn, "Unexpected exeception attempting to delete on undo '{0}': {1}", _dest, ex.Message);
                     }
 
                     // restore
@@ -108,7 +108,7 @@ namespace ClearCanvas.Dicom.Utilities.Command
                     }
                     catch (Exception ex)
                     {
-                        Platform.Log(LogLevel.Warn, "Error occurred when rolling back CopyDirectoryCommand: {0}", ex);
+                        Platform.Log(LogLevel.Warn, "Error occurred when rolling back CopyDirectoryCommand: {0}", ex.Message);
                     }
                 }
             }
@@ -138,12 +138,13 @@ namespace ClearCanvas.Dicom.Utilities.Command
         {
             try
             {
-                DirectoryUtility.DeleteIfExists(_backupDestDir);
+				if (!string.IsNullOrEmpty(_backupDestDir))
+					DirectoryUtility.DeleteIfExists(_backupDestDir);
             }
-            catch
+            catch (Exception x)
             {
             	//ignore
-                Platform.Log(LogLevel.Warn, "Unexpected exeception attempting to delete: {0}", _dest);
+				Platform.Log(LogLevel.Warn, "Unexpected exeception attempting to delete backup directory '{0}': {1}", _backupDestDir, x.Message);
             }
         }
